rest_api_server/requirements.txt
django 
djangorestframework 
django-cors-headers 
django_pg8000 
grpcio  
grpcio-tools 
lxml